next permutation
Find the point a[i]<a[i+1] by iterating from last. This is our index1.
If there is no breakpoint that means the array is in descending order. So, simply reverse the full array.
Else again iterate from last and find the element that is great than a[index1]. Store it as index2.
Swap the 2elements at index1 and index2.
Finally reverse the array from index1+1, so that we get the minimum number. Which means the next permutation.
Time complexity: O(n).